# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: VelumX MPC API
  version: 0.1.0
  description: |
    Stacks-native embedded wallet infrastructure. Like Privy for Stacks.

    ## Authentication
    All protected endpoints require an API key passed in the `x-api-key` header.
    Get your API key by registering at [velumx.xyz](https://velumx.xyz).

    ## Rate Limits
    - 100 requests/minute per API key
    - 10 requests/second burst

    ## Error Codes
    | Code | Meaning |
    |------|---------|
    | 400  | Bad request — invalid parameters |
    | 401  | Unauthorized — missing or invalid API key |
    | 404  | Not found — resource doesn't exist |
    | 409  | Conflict — resource already exists |
    | 429  | Too many requests — rate limit exceeded |
    | 500  | Internal server error |

  contact:
    name: VelumX
    url: https://velumx.xyz
  license:
    name: Proprietary
servers:
  - url: https://api.velumx.xyz/api/v1
    description: Production
  - url: http://localhost:4002/api/v1
    description: Local development
security:
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Your VelumX API key. Get one at https://velumx.xyz/dashboard.
  schemas:
...
    CreateWalletRequest:
      type: object
      required: [stxAddress, publicKey, userId]
      properties:
        stxAddress:
          type: string
          pattern: '^S[TP][A-Z0-9]{38,40}$'
          description: Stacks address
          example: SP2ZNGJ85WSY6PQTSKWDY0MM7NKJQZ5B8N9Y8W6P
        publicKey:
          type: string
          description: Compressed secp256k1 public key (hex)
          example: 02a17b...
        userId:
          type: string
          description: Your application's end-user ID
          example: user_abc123
        label:
          type: string
          maxLength: 100
          description: Human-readable label
        turnkeyWalletId:
          type: string
        derivationPath:
          type: string
          description: BIP32 derivation path
          example: m/44'/5757'/0'/0/0
        network:
          type: string
          enum: [mainnet, testnet]
          default: mainnet
    UpdateWalletRequest:
      type: object
      properties:
        label:
          type: string
          maxLength: 100
    MpcWallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        userId:
          type: string
        label:
          type: string
          nullable: true
        stxAddress:
          type: string
        publicKey:
          type: string
        turnkeyWalletId:
          type: string
        derivationPath:
          type: string
          nullable: true
        network:
          type: string
          enum: [mainnet, testnet]
        nonce:
          type: integer
        isActive:
          type: boolean
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedWallets:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MpcWallet'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...
    WalletExport:
      type: object
      properties:
        walletId:
          type: string
          format: uuid
        exportedAt:
          type: string
          format: date-time
        shareCount:
          type: integer
        shares:
          type: array
          items:
            type: object
            properties:
              shareIndex:
                type: integer
              holderId:
                type: string
                enum: [client, server-s1, server-s2]
              encryptedShare:
                type: string
                description: AES-256-GCM encrypted + base64 encoded
              encryptionKeyId:
                type: string
              shareVersion:
                type: integer
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
paths:
...
  /wallets:
    get:
      tags: [Wallets]
      summary: List wallets
      description: Paginated list of all wallets in your organization.
      operationId: listWallets
      parameters:
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by end-user ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated wallet list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWallets'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Wallets]
      summary: Create a new wallet
      description: Register an MPC wallet for an end-user. The wallet must already exist in Turnkey.
      operationId: createWallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MpcWallet'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: MPC not provisioned — call /auth/mpc/provision first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Wallet with this address already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wallets/{id}:
    get:
      tags: [Wallets]
      summary: Get wallet by ID
      operationId: getWallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Wallet details with balances and recent transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MpcWallet'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Wallets]
      summary: Update wallet label
      operationId: updateWallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWalletRequest'
      responses:
        '200':
          description: Wallet updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MpcWallet'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Wallets]
      summary: Soft-delete wallet
      description: Marks wallet as deleted. 7-day cooldown before permanent deletion.
      operationId: deleteWallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Wallet deleted
        '400':
          description: Wallet already deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /wallets/{id}/export:
    post:
      tags: [Wallets]
      summary: Export wallet key shares
      description: Export the MPC key shares for backup or migration.
      operationId: exportWallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Wallet key shares exported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /wallets/address/{address}:
    get:
      tags: [Wallets]
      summary: Get wallet by Stacks address
      operationId: getWalletByAddress
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
            pattern: '^S[TP][A-Z0-9]{38,40}$'
      responses:
        '200':
          description: Wallet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MpcWallet'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /wallets/user/{userId}:
    get:
      tags: [Wallets]
      summary: Get wallets by end-user ID
      operationId: getWalletsByUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of user's wallets (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MpcWallet'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            message: Unauthorized
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: Not found
...